# openapi.cloudflare.yaml
openapi: 3.0.3
info:
  title: ExitRoute API
  description: A self-hosted informational API. It never authenticates to, operates, or stores credentials
    for a consumer account.
  version: 0.1.0
paths:
...
  /v1/services:
    get:
      tags:
      - routes
      summary: List Services
      operationId: listServices
      security:
      - ApiKeyAuth: []
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          title: Cursor
          type: string
          minLength: 1
          maxLength: 512
          nullable: true
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 25
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServicePage'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /v1/observations:
    post:
      tags:
      - observations
      summary: Create Observation
      operationId: createObservation
      security:
      - ApiKeyAuth: []
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 16
          maxLength: 128
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObservationCreate'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservationReceipt'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /admin/v1/services:
    post:
      tags:
      - admin
      summary: Admin Create Service
      operationId: adminCreateService
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceView'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '409':
          $ref: '#/components/responses/Conflict'
      security:
      - BootstrapAdminAuth: []
      - ApiKeyAuth: []
  /admin/v1/routes:
    post:
      tags:
      - admin
      summary: Admin Create Route
      operationId: adminCreateRoute
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteView'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '409':
          $ref: '#/components/responses/Conflict'
      security:
      - BootstrapAdminAuth: []
      - ApiKeyAuth: []
  /admin/v1/revisions:
    post:
      tags:
      - admin
      summary: Admin Create Revision
      operationId: adminCreateRevision
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevisionCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevisionDraftView'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '409':
          $ref: '#/components/responses/Conflict'
      security:
      - BootstrapAdminAuth: []
      - ApiKeyAuth: []
  /admin/v1/revisions/{revision_id}/verifications:
    post:
      tags:
      - admin
      summary: Admin Verify Revision
      operationId: adminVerifyRevision
      security:
      - BootstrapAdminAuth: []
      - ApiKeyAuth: []
      parameters:
      - name: revision_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Revision Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationView'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/v1/revisions/{revision_id}/publish:
    post:
      tags:
      - admin
      summary: Admin Publish Revision
      operationId: adminPublishRevision
      security:
      - BootstrapAdminAuth: []
      - ApiKeyAuth: []
      parameters:
      - name: revision_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Revision Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishRevision'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevisionStateView'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/v1/revisions/{revision_id}/mark-stale:
    post:
      tags:
      - admin
      summary: Admin Mark Stale
      operationId: adminMarkRevisionStale
      security:
      - BootstrapAdminAuth: []
      - ApiKeyAuth: []
      parameters:
      - name: revision_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Revision Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevisionStateView'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/v1/revisions/{revision_id}/withdraw:
    post:
      tags:
      - admin
      summary: Admin Withdraw Revision
      operationId: adminWithdrawRevision
      security:
      - BootstrapAdminAuth: []
      - ApiKeyAuth: []
      parameters:
      - name: revision_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Revision Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevisionStateView'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/v1/api-clients:
    post:
      tags:
      - admin
      summary: Admin Create Client
      operationId: adminCreateApiClient
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiClientCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiClientView'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '409':
          $ref: '#/components/responses/Conflict'
      security:
      - BootstrapAdminAuth: []
      - ApiKeyAuth: []
  /admin/v1/api-clients/{client_id}/keys:
    post:
      tags:
      - admin
      summary: Admin Create Key
      operationId: adminCreateApiKey
      security:
      - BootstrapAdminAuth: []
      - ApiKeyAuth: []
      parameters:
      - name: client_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Client Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreated'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      tags:
      - admin
      summary: Admin List Keys
      operationId: adminListApiKeys
      security:
      - BootstrapAdminAuth: []
      - ApiKeyAuth: []
      parameters:
      - name: client_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Client Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKeyView'
                title: Response Adminlistapikeys
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/v1/api-keys/{key_id}/rotate:
    post:
      tags:
      - admin
      summary: Admin Rotate Key
      operationId: adminRotateApiKey
      security:
      - BootstrapAdminAuth: []
      - ApiKeyAuth: []
      parameters:
      - name: key_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Key Id
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreated'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/v1/api-keys/{key_id}:
    delete:
      tags:
      - admin
      summary: Admin Revoke Key
      operationId: adminRevokeApiKey
      security:
      - BootstrapAdminAuth: []
      - ApiKeyAuth: []
      parameters:
      - name: key_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Key Id
      responses:
        '204':
          description: Successful Response
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/v1/api-clients/{client_id}:
    delete:
      tags:
      - admin
      summary: Admin Deactivate Client
      operationId: adminDeactivateApiClient
      security:
      - BootstrapAdminAuth: []
      - ApiKeyAuth: []
      parameters:
      - name: client_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Client Id
      responses:
        '204':
          description: Successful Response
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/v1/observations:
    get:
      tags:
      - admin
      summary: Admin List Observations
      operationId: adminListObservations
      security:
      - BootstrapAdminAuth: []
      - ApiKeyAuth: []
      parameters:
      - name: moderation_state
        in: query
        required: false
        schema:
          enum:
          - queued
          - accepted
          - rejected
          type: string
          default: queued
          title: Moderation State
      - name: cursor
        in: query
        required: false
        schema:
          title: Cursor
          type: string
          minLength: 1
          maxLength: 512
          nullable: true
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 25
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservationAdminPage'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /admin/v1/observations/{observation_id}/moderate:
    post:
      tags:
      - admin
      summary: Admin Moderate Observation
      operationId: adminModerateObservation
      security:
      - BootstrapAdminAuth: []
      - ApiKeyAuth: []
      parameters:
      - name: observation_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Observation Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObservationModerate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservationAdminView'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
components:
  schemas:
...
    ObservationReceipt:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        moderation_state:
          type: string
          enum:
          - queued
          - accepted
          - rejected
          title: Moderation State
          default: queued
        duplicate:
          type: boolean
          title: Duplicate
        received_at:
          type: string
          format: date-time
          title: Received At
      additionalProperties: false
      type: object
      required:
      - id
      - duplicate
      - received_at
      title: ObservationReceipt
...
    ServicePage:
      properties:
        data:
          items:
            $ref: '#/components/schemas/ServiceView'
          type: array
          title: Data
        next_cursor:
          title: Next Cursor
          type: string
          nullable: true
      additionalProperties: false
      type: object
      required:
      - data
      - next_cursor
      title: ServicePage
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      description: A scoped key created by an ExitRoute administrator.
      in: header
      name: X-API-Key
    BootstrapAdminAuth:
      type: http
      description: Bootstrap-only bearer token. Prefer a scoped admin API key after setup.
      scheme: bearer
  responses:
...
    Unauthorized:
      description: Authentication failed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: The credential lacks the required scope.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: The requested resource does not exist.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: The request conflicts with resource state.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unprocessable:
      description: The request violates a semantic or safety rule.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    RateLimited:
      description: The API-client quota was exceeded.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        Retry-After:
          description: Seconds until a new quota window.
          schema:
            type: integer
            minimum: 1
    InternalError:
      description: The server could not complete the request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...